Готовое ДЗ Props, ref, onMounted с учетом корректировок
