+++ src/assets/main.css
@@ -816,16 +816,12 @@ body {
   background-color: #eaeef6;
 }
 .main__block {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-around;
   width: 100%;
   margin: 0 auto;
   padding: 25px 0 49px;
 }
 .main__content {
   width: 100%;
-  display: flex;
 }
 .main__column {
   width: 20%;
